requirements.txt
alabaster==0.7.12
appdirs==1.4.4
atomicwrites==1.4.0
attrs==21.4.0
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.10
colorama==0.4.4
commonmark==0.9.1
coverage==6.3.2
docutils==0.17.1
flake8==4.0.1
idna==3.3
imagesize==1.3.0
importlib-metadata==4.11.3
iniconfig==1.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
mccabe==0.6.1
packaging==21.3
playsound==1.3.0
pluggy==1.0.0
py==1.11.0
PyAudio-0.2.11-cp39-cp39-win_amd64.whl
pycodestyle==2.8.0
pydub==0.25.1
pyflakes==2.4.0
Pygments==2.11.2
pyparsing==3.0.7
pytest==7.1.1
pytest-cov==3.0.0
pytz==2021.3
recommonmark==0.7.1
requests==2.27.1
rinoh-typeface-dejavuserif==0.1.3
rinoh-typeface-texgyrecursor==0.1.1
rinoh-typeface-texgyreheros==0.1.1
rinoh-typeface-texgyrepagella==0.1.1
rinohtype==0.5.3
snowballstemmer==2.2.0
Sphinx==4.4.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
tomli==2.0.1
urllib3==1.26.8
zipp==3.7.0
